Fix parsing the string into long value

+++ language/src/main/java/gh/marad/chi/language/nodes/expr/cast/CastToLongExpr.java
@@ -24,7 +24,7 @@ public abstract class CastToLongExpr extends CastExpression {
     @Specialization
     @CompilerDirectives.TruffleBoundary
     long fromString(String value) {
-        return Integer.parseInt(value);
+        return Long.parseLong(value);
     }
 
     @Override
